refactor(api): add decision surface runner
