+++ crates/api/Cargo.toml
@@ -26,6 +26,7 @@ rustc-hash = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 schemars = { workspace = true, optional = true }
+xxhash-rust = { workspace = true }
 
 [features]
 schema = ["dep:schemars", "fallow-engine/schema", "fallow-output/schema", "fallow-types/schema"]
