+++ crates/api/src/runtime/mod.rs
@@ -8,6 +8,7 @@ use fallow_types::workspace::WorkspaceDiagnostic;
 
 mod audit;
 mod dead_code;
+mod decision_surface;
 mod duplication;
 mod feature_flags;
 mod trace;
@@ -16,14 +17,16 @@ pub use crate::runtime_output::{
     AuditProgrammaticKeySnapshot, AuditProgrammaticOutput, BoundaryViolationsOutput,
     BoundaryViolationsProgrammaticOutput, CircularDependenciesOutput,
     CircularDependenciesProgrammaticOutput, DeadCodeOutput, DeadCodeProgrammaticOutput,
-    DuplicationOutput, DuplicationProgrammaticOutput, FeatureFlagsOutput,
-    FeatureFlagsProgrammaticOutput, HealthJsonReportInput, HealthProgrammaticOutput,
-    TraceCloneOutput, TraceCloneProgrammaticOutput, TraceDependencyOutput,
-    TraceDependencyProgrammaticOutput, TraceExportOutput, TraceExportProgrammaticOutput,
-    TraceFileOutput, TraceFileProgrammaticOutput, serialize_health_report_json,
+    DecisionSurfaceProgrammaticOutput, DuplicationOutput, DuplicationProgrammaticOutput,
+    FeatureFlagsOutput, FeatureFlagsProgrammaticOutput, HealthJsonReportInput,
+    HealthProgrammaticOutput, TraceCloneOutput, TraceCloneProgrammaticOutput,
+    TraceDependencyOutput, TraceDependencyProgrammaticOutput, TraceExportOutput,
+    TraceExportProgrammaticOutput, TraceFileOutput, TraceFileProgrammaticOutput,
+    serialize_health_report_json,
 };
 pub use audit::run_audit;
 pub use dead_code::{run_boundary_violations, run_circular_dependencies, run_dead_code};
+pub use decision_surface::run_decision_surface;
 pub use duplication::run_duplication;
 pub use feature_flags::run_feature_flags;
 pub use trace::{run_trace_clone, run_trace_dependency, run_trace_export, run_trace_file};
